refactored main menu and footer retrival into context
