+++ CMS/context_processors/__init__.py
@@ -1,0 +1,1 @@
+from .nav_context_processor import nav_menu_render
