+++ CMS/static/scss/partials/nav.scss
@@ -308,7 +308,7 @@
         display: flex;
         justify-content: space-between;
         list-style: none;
-        padding: 0px 15px 0px 0px;
+        padding: 0px 20px 0px 0px;
         margin: 5px 5px 5px 30px;
         background-color: #000000;
         border-radius: 40px 40px 40px 40px;
